fix(material): normals can be negatives, no mirror on normals

+++ src/material.hh
@@ -21,7 +21,7 @@ public:
     virtual ~TextureMaterial() = default;
 
     virtual LocalTexture get_texture_at(const Point3 &p) const = 0;
-    virtual Vector3 get_normal_at(float y, float x) const = 0;
+    virtual Vector3 get_normal_at(double y, double x) const = 0;
 };
 
 class UniformTexture : public TextureMaterial
@@ -33,7 +33,7 @@ public:
     UniformTexture(LocalTexture tex);
 
     LocalTexture get_texture_at(const Point3 &p) const override;
-    Vector3 get_normal_at(float y, float x) const override;
+    Vector3 get_normal_at(double y, double x) const override;
 
     const static UniformTexture default_mat;
 };
@@ -52,5 +52,5 @@ public:
                  Vector3 normal_scale = Vector3(1.0, 1.0, 1.0));
 
     LocalTexture get_texture_at(const Point3 &p) const override;
-    Vector3 get_normal_at(float y, float x) const override;
+    Vector3 get_normal_at(double y, double x) const override;
 };
